requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-openai
pinecone-client
gradio
PyGithub
python-dotenv
